Add platform media layer
