+++ Demo/exception_info/platform_media.c
@@ -1,0 +1,207 @@
+
+#include "platform_media.h"
+
+#include "stm32h7xx_hal.h"
+
+static uint32_t GetSector(uint32_t Address);
+
+/*-----------------------------------------------------------*/
+
+/**
+  * @brief  Unlocks Flash for write access
+  * @param  None
+  * @retval None
+  */
+static void FLASH_If_Init(void)
+{
+  
+  /* Clear pending flags (if any) */  
+  __HAL_FLASH_CLEAR_FLAG(FLASH_FLAG_EOP | FLASH_FLAG_OPERR | FLASH_FLAG_WRPERR | 
+                         FLASH_FLAG_PGSERR | FLASH_FLAG_WRPERR);
+}
+/*-----------------------------------------------------------*/
+
+/**
+  * @brief  This function does an erase of all user flash area
+  * @param  StartSector: start of user flash area
+  * @retval 0: user flash area successfully erased
+  *         1: error occurred
+  */
+uint32_t FLASH_If_Erase(uint32_t StartSector)
+{
+  uint32_t UserStartSector;
+  uint32_t SectorError;
+  FLASH_EraseInitTypeDef pEraseInit;
+
+  /* Unlock the Flash to enable the flash control register access *************/ 
+  HAL_FLASH_Unlock(); 
+  FLASH_If_Init();
+  
+  /* Get the sector where start the user flash area */
+  UserStartSector = GetSector(APPLICATION_ADDRESS);
+  
+  pEraseInit.TypeErase = FLASH_TYPEERASE_SECTORS;
+  pEraseInit.Sector = UserStartSector;
+  pEraseInit.NbSectors = 8 - UserStartSector;
+  pEraseInit.VoltageRange = FLASH_VOLTAGE_RANGE_3;
+  
+  if (APPLICATION_ADDRESS < ADDR_FLASH_SECTOR_0_BANK2)
+  {
+    pEraseInit.Banks = FLASH_BANK_1;
+    if (HAL_FLASHEx_Erase(&pEraseInit, &SectorError) != HAL_OK)
+    {
+      /* Error occurred while sector erase */
+      return (1);
+    }
+    
+    /* Mass erase of second bank */
+    pEraseInit.TypeErase = FLASH_TYPEERASE_MASSERASE;
+    pEraseInit.Banks = FLASH_BANK_2;
+    if (HAL_FLASHEx_Erase(&pEraseInit, &SectorError) != HAL_OK)
+    {
+      /* Error occurred while sector erase */
+      return (1);
+    }
+  }
+  else
+  {
+    pEraseInit.TypeErase = FLASH_TYPEERASE_SECTORS;
+    pEraseInit.Banks = FLASH_BANK_2;
+    if (HAL_FLASHEx_Erase(&pEraseInit, &SectorError) != HAL_OK)
+    {
+      /* Error occurred while sector erase */
+      return (1);
+    }
+  }
+  HAL_FLASH_Lock(); 
+  return (0);
+}
+/*-----------------------------------------------------------*/
+
+/**
+  * @brief  Gets the sector of a given address
+  * @param  Address: Flash address
+  * @retval The sector of a given address
+  */
+static uint32_t GetSector(uint32_t Address)
+{
+  uint32_t sector = 0;
+  
+  if((Address < ADDR_FLASH_SECTOR_1_BANK1) && (Address >= ADDR_FLASH_SECTOR_0_BANK1))
+  {
+    sector = FLASH_SECTOR_0;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_2_BANK1) && (Address >= ADDR_FLASH_SECTOR_1_BANK1))
+  {
+    sector = FLASH_SECTOR_1;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_3_BANK1) && (Address >= ADDR_FLASH_SECTOR_2_BANK1))
+  {
+    sector = FLASH_SECTOR_2;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_4_BANK1) && (Address >= ADDR_FLASH_SECTOR_3_BANK1))
+  {
+    sector = FLASH_SECTOR_3;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_5_BANK1) && (Address >= ADDR_FLASH_SECTOR_4_BANK1))
+  {
+    sector = FLASH_SECTOR_4;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_6_BANK1) && (Address >= ADDR_FLASH_SECTOR_5_BANK1))
+  {
+    sector = FLASH_SECTOR_5;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_7_BANK1) && (Address >= ADDR_FLASH_SECTOR_6_BANK1))
+  {
+    sector = FLASH_SECTOR_6;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_0_BANK2) && (Address >= ADDR_FLASH_SECTOR_7_BANK1))
+  {
+    sector = FLASH_SECTOR_7;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_1_BANK2) && (Address >= ADDR_FLASH_SECTOR_0_BANK2))
+  {
+    sector = FLASH_SECTOR_0;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_2_BANK2) && (Address >= ADDR_FLASH_SECTOR_1_BANK2))
+  {
+    sector = FLASH_SECTOR_1;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_3_BANK2) && (Address >= ADDR_FLASH_SECTOR_2_BANK2))
+  {
+    sector = FLASH_SECTOR_2;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_4_BANK2) && (Address >= ADDR_FLASH_SECTOR_3_BANK2))
+  {
+    sector = FLASH_SECTOR_3;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_5_BANK2) && (Address >= ADDR_FLASH_SECTOR_4_BANK2))
+  {
+    sector = FLASH_SECTOR_4;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_6_BANK2) && (Address >= ADDR_FLASH_SECTOR_5_BANK2))
+  {
+    sector = FLASH_SECTOR_5;  
+  }
+  else if((Address < ADDR_FLASH_SECTOR_7_BANK2) && (Address >= ADDR_FLASH_SECTOR_6_BANK2))
+  {
+    sector = FLASH_SECTOR_6;  
+  }
+  else /*if((Address < USER_FLASH_END_ADDRESS) && (Address >= ADDR_FLASH_SECTOR_7_BANK2))*/
+  {
+    sector = FLASH_SECTOR_7;  
+  }
+
+  return sector;
+}
+/*-----------------------------------------------------------*/
+
+/**
+  * @brief  This function writes a data buffer in flash (data are 32-bit aligned).
+  * @note   After writing data buffer, the flash content is checked.
+  * @param  FlashAddress: start address for writing data buffer
+  * @param  Data: pointer on data buffer
+  * @param  DataLength: length of data buffer (unit is 32-bit word)   
+  * @param  pNextFlashAddress: Last flash address to write
+  * @retval 0: Data successfully written to Flash memory
+  *         1: Error occurred while writing data in Flash memory
+  *         2: Written Data in flash memory is different from expected one
+  */
+uint32_t FLASH_If_Write(uint32_t FlashAddress, uint32_t* Data ,uint32_t DataLength, uint32_t *pNextFlashAddress )
+{
+  uint32_t i = 0;
+ /* Unlock the Flash to enable the flash control register access *************/ 
+  HAL_FLASH_Unlock();
+
+  for (i = 0; (i < DataLength) && (FlashAddress <= (USER_FLASH_END_ADDRESS-32)); i+=8)
+  {
+    /* Device voltage range supposed to be [2.7V to 3.6V], the operation will
+       be done by word */ 
+    if (HAL_FLASH_Program(FLASH_TYPEPROGRAM_FLASHWORD, FlashAddress, ((uint32_t)(Data+i))) == HAL_OK)
+    {
+     /* Check the written value */
+      if (*(uint32_t*)FlashAddress != *(uint32_t*)(Data+i))
+      {
+        /* Flash content doesn't match SRAM content */
+        return(FLASHIF_WRITINGCTRL_ERROR);
+      }
+      /* Increment FLASH destination address */
+      FlashAddress += 32;
+    }
+    else
+    {
+      /* Error occurred while writing data in Flash memory */
+      return (FLASHIF_WRITING_ERROR);
+    }
+  }
+
+  HAL_FLASH_Lock();
+
+  if( pNextFlashAddress != NULL )
+  {
+    *pNextFlashAddress = FlashAddress;
+  }
+
+  return (FLASHIF_OK);
+}
+/*-----------------------------------------------------------*/
